WIP: Doesn't treat truncated alleles as contaminations
Some alleles are truncations of other alleles.  Because both matched, they were
treated as contamination.  This checks whether one match is a subset of another
and doesn't mark it as a contamination if this is the case.

I've spotted some bugs / features in the original algorithm which I have marked
but not fixed.

One of the old tests is now wrong which I will address in a subsequent commit.